// frontend/src/company.d.ts
export interface CompanyKeyRatios {
  symbol: string;
  grossProfitMarginTTM: number;
  ebitMarginTTM: number;
  ebitdaMarginTTM: number;
  operatingProfitMarginTTM: number;
  pretaxProfitMarginTTM: number;
  continuousOperationsProfitMarginTTM: number;
  netProfitMarginTTM: number;
  bottomLineProfitMarginTTM: number;
  receivablesTurnoverTTM: number;
  payablesTurnoverTTM: number;
  inventoryTurnoverTTM: number;
  fixedAssetTurnoverTTM: number;
  assetTurnoverTTM: number;
  currentRatioTTM: number;
  quickRatioTTM: number;
  solvencyRatioTTM: number;
  cashRatioTTM: number;
  priceToEarningsRatioTTM: number;
  priceToEarningsGrowthRatioTTM: number;
  forwardPriceToEarningsGrowthRatioTTM: number;
  priceToBookRatioTTM: number;
  priceToSalesRatioTTM: number;
  priceToFreeCashFlowRatioTTM: number;
  priceToOperatingCashFlowRatioTTM: number;
  debtToAssetsRatioTTM: number;
  debtToEquityRatioTTM: number;
  debtToCapitalRatioTTM: number;
  longTermDebtToCapitalRatioTTM: number;
  financialLeverageRatioTTM: number;
  workingCapitalTurnoverRatioTTM: number;
  operatingCashFlowRatioTTM: number;
  operatingCashFlowSalesRatioTTM: number;
  freeCashFlowOperatingCashFlowRatioTTM: number;
  debtServiceCoverageRatioTTM: number;
  interestCoverageRatioTTM: number;
  shortTermOperatingCashFlowCoverageRatioTTM: number;
  operatingCashFlowCoverageRatioTTM: number;
  capitalExpenditureCoverageRatioTTM: number;
  dividendPaidAndCapexCoverageRatioTTM: number;
  dividendPayoutRatioTTM: number;
  dividendYieldTTM: number;
  enterpriseValueTTM: number;
  revenuePerShareTTM: number;
  netIncomePerShareTTM: number;
  interestDebtPerShareTTM: number;
  cashPerShareTTM: number;
  bookValuePerShareTTM: number;
  tangibleBookValuePerShareTTM: number;
  shareholdersEquityPerShareTTM: number;
  operatingCashFlowPerShareTTM: number;
  capexPerShareTTM: number;
  freeCashFlowPerShareTTM: number;
  netIncomePerEBTTTM: number;
  ebtPerEbitTTM: number;
  priceToFairValueTTM: number;
  debtToMarketCapTTM: number;
  effectiveTaxRateTTM: number;
  enterpriseValueMultipleTTM: number;
}
// ...
export interface CompanyIncomeStatement {
  date: string;
  symbol: string;
  reportedCurrency: string;
  cik: string;
  fillingDate: string;
  acceptedDate: string;
  calendarYear: string;
  period: string;
  revenue: number;
  costOfRevenue: number;
  grossProfit: number;
  grossProfitRatio: number;
  researchAndDevelopmentExpenses: number;
  generalAndAdministrativeExpenses: number;
  sellingAndMarketingExpenses: number;
  sellingGeneralAndAdministrativeExpenses: number;
  otherExpenses: number;
  operatingExpenses: number;
  costAndExpenses: number;
  interestIncome: number;
  interestExpense: number;
  depreciationAndAmortization: number;
  ebitda: number;
  ebitdaratio: number;
  operatingIncome: number;
  operatingIncomeRatio: number;
  totalOtherIncomeExpensesNet: number;
  incomeBeforeTax: number;
  incomeBeforeTaxRatio: number;
  incomeTaxExpense: number;
  netIncome: number;
  netIncomeRatio: number;
  eps: number;
  epsdiluted: number;
  weightedAverageShsOut: number;
  weightedAverageShsOutDil: number;
  link: string;
  finalLink: string;
}